Implements patch update procedure from Morris’s "Contagion"
From the Info tab:

[This is a slightly hacked version of the NetLogo models lib Rumor Mill.
This model implements the patch update procedure from Stephen Morris’s
"Contagion", The Review of Economic Studies 2000 67(1) 57-78, as
presented in Vega-Redondo 2007. This means that a patch acquires the
"rumor" if and only if the proportion of its neighbors (Moore 8 or von
Neuman 4) who have the rumor already is greater than a parameter q. All
of the documentation not in square brackets is from the original Rumor
Mill model. -Marshall Abrams]